requirements.txt
python-dotenv~=1.0.1
pip~=21.2.3
pytz~=2024.1
numpy~=1.26.2
Cython~=3.0.10
setuptools~=57.4.0
click~=8.1.7
pandas~=2.2.1
matplotlib~=3.8.4
scipy~=1.13.0
python-dateutil~=2.9.0.post0
six~=1.16.0